(refactor) Make a module for solver methods
